Corrigido erro de token expirado no middleware, criado verificações nos Services

+++ src/services/FavoritesCreateServices.js
@@ -1,9 +1,19 @@
+const AppError = require("../utils/AppError");
+
 class FavoritesCreateServices {
     constructor(favoritesRepository) {
         this.favoritesRepository = favoritesRepository;
     }
 
     async create({ dish_id, user_id, dish_title }) {
+        if(!dish_id){
+            throw new AppError("Id do prato não encontrado", 400)
+        }
+        
+        if(!dish_title){
+            throw new AppError("Titulo do prato não encontrado", 400)
+        }
+
         const favorite = this.favoritesRepository.createFavorites({ dish_id, user_id, dish_title })
 
         return favorite
@@ -16,6 +26,9 @@ class FavoritesCreateServices {
     }
 
     async delete({ favorite_id }) {
+        if(!favorite_id){
+            throw new AppError("Id não encontrado")
+        }
         this.favoritesRepository.DeleteFavorites({ favorite_id })
     }
 }
